CRUD implementation for products
feat: CRUD implementation for products

+++ Mealthy/Mealthy/Resource/ProductResource.cs
@@ -5,7 +5,7 @@ public class ProductResource
     public int Id { get; set; }
     public string Name { get; set; }
     public string Category { get; set; }
-    public int Price { get; set; }
+    public float Price { get; set; }
     public string Unit { get; set; }
     public int Quantity { get; set; }
     public string photoUrl { get; set; }
